fix bug: memory reallocate maybe leak
git-svn-id: http://192.168.20.11/sequoiadb/trunk@37857 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/tools/inspect/sdbInspect.cpp
@@ -1339,12 +1339,12 @@ INT32 readCiRecord( OSSFILE &in, INT64 &offset,
 {
    INT32 rc         = SDB_OK ;
    CHAR *bsonBuffer = NULL ;
-   INT32 bufferLen  = 0 ;
+   INT64 bufferLen  = 0 ;
 
    UINT32 idx = 0 ;
    while ( idx < header._recordCount )
    {
-      INT32 recordLen = 0 ;
+      INT64 recordLen = 0 ;
       CHAR  state ;
       rc = readFromFile( in, offset, (CHAR *)&recordLen, sizeof( INT32 ) ) ;
       CHECK_VALUE( ( SDB_OK != rc ), error ) ;
